# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Inventory management API for doors and windows company
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Category {
  id: number;
  name: string;
  description?: string | null;
  color: string;
  createdAt: string;
}

export interface CreateCategoryInput {
  name: string;
  description?: string | null;
  color: string;
}

export interface Product {
  id: number;
  name: string;
  description?: string | null;
  categoryId: number;
  categoryName: string;
  categoryColor: string;
  unit: string;
  currentStock: number;
  minimumStock: number;
  createdAt: string;
}

export interface CreateProductInput {
  name: string;
  description?: string | null;
  categoryId: number;
  unit: string;
  currentStock: number;
  minimumStock: number;
}

export type MovementType = (typeof MovementType)[keyof typeof MovementType];

export const MovementType = {
  entrada: "entrada",
  salida: "salida",
} as const;

export interface Movement {
  id: number;
  productId: number;
  productName: string;
  categoryName: string;
  type: MovementType;
  quantity: number;
  notes?: string | null;
  stockBefore: number;
  stockAfter: number;
  createdAt: string;
}

export type CreateMovementInputType =
  (typeof CreateMovementInputType)[keyof typeof CreateMovementInputType];

export const CreateMovementInputType = {
  entrada: "entrada",
  salida: "salida",
} as const;

export interface CreateMovementInput {
  productId: number;
  type: CreateMovementInputType;
  quantity: number;
  notes?: string | null;
}

export interface DashboardSummary {
  totalProducts: number;
  totalCategories: number;
  lowStockProducts: number;
  recentMovements: Movement[];
  totalEntradasHoy: number;
  totalSalidasHoy: number;
}

export interface DeleteResult {
  success: boolean;
}

export type ListProductsParams = {
  categoryId?: number;
};

export type ListMovementsParams = {
  productId?: number;
  type?: ListMovementsType;
  limit?: number;
};

export type ListMovementsType =
  (typeof ListMovementsType)[keyof typeof ListMovementsType];

export const ListMovementsType = {
  entrada: "entrada",
  salida: "salida",
} as const;