# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Inventory management API for doors and windows company
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: categories
    description: Product categories
  - name: products
    description: Inventory products
  - name: movements
    description: Stock movements (in/out)
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List all categories
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [categories]
      summary: Create a category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryInput"
      responses:
        "201":
          description: Created category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"

  /categories/{id}:
    put:
      operationId: updateCategory
      tags: [categories]
      summary: Update a category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryInput"
      responses:
        "200":
          description: Updated category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
    delete:
      operationId: deleteCategory
      tags: [categories]
      summary: Delete a category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResult"

  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List all products
      parameters:
        - name: categoryId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductInput"
      responses:
        "201":
          description: Created product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"

  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    put:
      operationId: updateProduct
      tags: [products]
      summary: Update a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductInput"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResult"

  /movements:
    get:
      operationId: listMovements
      tags: [movements]
      summary: List all movements
      parameters:
        - name: productId
          in: query
          required: false
          schema:
            type: integer
        - name: type
          in: query
          required: false
          schema:
            type: string
            enum: [entrada, salida]
        - name: limit
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of movements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Movement"
    post:
      operationId: createMovement
      tags: [movements]
      summary: Register a stock movement
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMovementInput"
      responses:
        "201":
          description: Created movement
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Movement"

  /dashboard:
    get:
      operationId: getDashboard
      tags: [products]
      summary: Get dashboard summary stats
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Category:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
        color:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - color
        - createdAt

    CreateCategoryInput:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
        color:
          type: string
      required:
        - name
        - color

    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
        categoryId:
          type: integer
        categoryName:
          type: string
        categoryColor:
          type: string
        unit:
          type: string
        currentStock:
          type: number
        minimumStock:
          type: number
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - categoryId
        - categoryName
        - categoryColor
        - unit
        - currentStock
        - minimumStock
        - createdAt

    CreateProductInput:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
        categoryId:
          type: integer
        unit:
          type: string
        currentStock:
          type: number
        minimumStock:
          type: number
      required:
        - name
        - categoryId
        - unit
        - currentStock
        - minimumStock

    Movement:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        categoryName:
          type: string
        type:
          type: string
          enum: [entrada, salida]
        quantity:
          type: number
        notes:
          type: string
          nullable: true
        stockBefore:
          type: number
        stockAfter:
          type: number
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - productId
        - productName
        - categoryName
        - type
        - quantity
        - stockBefore
        - stockAfter
        - createdAt

    CreateMovementInput:
      type: object
      properties:
        productId:
          type: integer
        type:
          type: string
          enum: [entrada, salida]
        quantity:
          type: number
        notes:
          type: string
          nullable: true
      required:
        - productId
        - type
        - quantity

    DashboardSummary:
      type: object
      properties:
        totalProducts:
          type: integer
        totalCategories:
          type: integer
        lowStockProducts:
          type: integer
        recentMovements:
          type: array
          items:
            $ref: "#/components/schemas/Movement"
        totalEntradasHoy:
          type: integer
        totalSalidasHoy:
          type: integer
      required:
        - totalProducts
        - totalCategories
        - lowStockProducts
        - recentMovements
        - totalEntradasHoy
        - totalSalidasHoy

    DeleteResult:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success