# lib/api-zod/src/generated/types/createCategoryInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Inventory management API for doors and windows company
 * OpenAPI spec version: 0.1.0
 */

export interface CreateCategoryInput {
  name: string;
  description?: string | null;
  color: string;
}