# lib/api-zod/src/generated/types/createMovementInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Inventory management API for doors and windows company
 * OpenAPI spec version: 0.1.0
 */
import type { CreateMovementInputType } from "./createMovementInputType";

export interface CreateMovementInput {
  productId: number;
  type: CreateMovementInputType;
  quantity: number;
  notes?: string | null;
}