# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Inventory management API for doors and windows company
 * OpenAPI spec version: 0.1.0
 */
import type { Movement } from "./movement";

export interface DashboardSummary {
  totalProducts: number;
  totalCategories: number;
  lowStockProducts: number;
  recentMovements: Movement[];
  totalEntradasHoy: number;
  totalSalidasHoy: number;
}