# lib/api-zod/src/generated/types/movement.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Inventory management API for doors and windows company
 * OpenAPI spec version: 0.1.0
 */
import type { MovementType } from "./movementType";

export interface Movement {
  id: number;
  productId: number;
  productName: string;
  categoryName: string;
  type: MovementType;
  quantity: number;
  notes?: string | null;
  stockBefore: number;
  stockAfter: number;
  createdAt: Date;
}