# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Inventory management API for doors and windows company
 * OpenAPI spec version: 0.1.0
 */

export interface Product {
  id: number;
  name: string;
  description?: string | null;
  categoryId: number;
  categoryName: string;
  categoryColor: string;
  unit: string;
  currentStock: number;
  minimumStock: number;
  createdAt: Date;
}